+++ __tests__/api.test.js
@@ -1,0 +1,12 @@
+// test('jest should work correctly', () => {
+//   expect(true).toBe(true);
+// });
+
+describe('Test the root path', () => {
+    test('It should response the GET method', (done) => {
+        request(app).get('/').then((response) => {
+            expect(response.statusCode).toBe(200);
+            done();
+        });
+    });
+});
