--- __tests__/routes.test.js
@@ -1,3 +1,0 @@
-test('jest should work correctly', () => {
-  expect(true).toBe(true);
-});
